requirements.txt
# Core LiveKit agent framework
livekit-agents>=1.2.0

# LiveKit plugins (VAD, turn detection, etc.)
livekit-plugins-silero>=0.1.5
livekit-plugins-turn-detector>=0.1.3
livekit-plugins-openai>=0.5.0

# Install all required packages for our voice sales agent

# Faster-Whisper (STT)
faster-whisper>=1.0.3

# Ollama Python client (LLM via qwen2.5)
ollama>=0.1.7

# Piper TTS requires subprocess calls, no pip package needed
# But we install soundfile for handling audio I/O
soundfile>=0.12.1

# General utilities
python-dotenv>=1.0.1
numpy>=1.26.4

# Optional: logging/metrics helpers
loguru>=0.7.2
livekit-agents
faster-whisper
ollama
piper-tts
livekit-agents
faster-whisper
piper-tts
ollama
python-dotenv
livekit-agents
faster-whisper
piper-tts
ollama
python-dotenv
livekit-agents>=1.2.0
faster-whisper>=1.0.0
piper-tts>=1.2.0
ollama>=0.1.0
python-dotenv>=1.0.1
# Core dependencies
livekit-agents>=0.5.0
faster-whisper>=1.0

# Audio handling (tempfiles, streaming, encoding/decoding)
soundfile>=0.12.1
numpy>=1.23

# Typing + context tools (standard libs, but pinned for safety)
typing-extensions>=4.9.0


